+++ lightsim2grid/newtonpf.py
@@ -9,8 +9,6 @@
 """
 This module provide a function that can serve as a base replacement to the function newtonpf of
 `pandapower/pypower/newtonpf.py`
-
-
 """
 import numpy as np
 from scipy import sparse
@@ -23,13 +21,23 @@ try:
 except ImportError:
     pass
 
+from pandapower.pypower.idx_bus import SL_FAC
 
-def newtonpf(Ybus, Sbus, V0, pv, pq, ppci, options):
+def newtonpf_old(Ybus, Sbus, V0, pv, pq, ppci, options):
     """
     Perform the Newton scheme to compute the AC powerflow of the system provided as input.
     It supports only one single slack bus.
 
     It is main as being integrated into pandapower as a replacement of the pypower implementation of "newtonpf"
+
+    .. versionadded:: 0.5.6
+
+    .. note::
+        This is a legacy code mainly present for compatibility with older pandapower versions.
+
+    .. warning::
+        It considers that all nodes non pv, non pq are slack nodes and assign the same weights
+        for all of them
 
     Parameters
     ----------
@@ -71,6 +79,100 @@ def newtonpf(Ybus, Sbus, V0, pv, pq, ppci, options):
     """
     max_it = options["max_iteration"]
     tol = options['tolerance_mva']
+
+    # initialize the solver
+    # TODO have that in options maybe (can use GaussSeidel, and NR with KLU -faster- or SparseLU)
+    if KLU_solver_available:
+        solver = KLUSolver()
+    else:
+        solver = SparseLUSolver()
+    Ybus = sparse.csc_matrix(Ybus)
+
+    # extract the slack bus
+    ref = set(np.arange(Sbus.shape[0])) - set(pv) - set(pq)
+    ref = np.array(list(ref))
+    # build the slack weights
+    slack_weights = np.zeros(Sbus.shape[0])
+    slack_weights[ref] = 1.0 / ref.shape[0]
+
+    # do the newton raphson algorithm
+    solver.solve(Ybus, V0, Sbus, ref, slack_weights, pv, pq, max_it, tol)
+
+    # extract the results
+    Va = solver.get_Va()
+    Vm = solver.get_Vm()
+    V = Vm * np.exp(1j * Va)
+    J = solver.get_J()
+    converged = solver.converged()
+    iterations = solver.get_nb_iter()
+
+    Vm_it = None
+    Va_it = None
+    return V, converged, iterations, J, Vm_it, Va_it
+
+
+def newtonpf(Ybus, Sbus, V0, ref, pv, pq, ppci, options):
+    """
+    Perform the Newton scheme to compute the AC powerflow of the system provided as input.
+    It supports only one single slack bus.
+
+    It is main as being integrated into pandapower as a replacement of the pypower implementation of "newtonpf"
+
+    .. versionchanged:: 0.5.6
+
+    .. note::
+        It has been updated in version 0.5.6 to match pandapower new signature (addition of the `ref`
+        parameter)
+
+        If you want the old behaviour, please use the `newtonpf_old`function.
+    
+    Parameters
+    ----------
+    Ybus: ``numpy.ndarray``, ``numpy.sparmatrix``, dtype:complex
+        The admittance matrix. If not in a sparse CSC format, it will be converted to it.
+
+    Sbus: ``numpy.ndarray``, dtype:complex
+        The power injected at each bus.
+
+    V0: ``numpy.ndarray``, dtype:complex
+        The initial voltage
+
+    ref:
+        Id of the slack buses (added in version 0.5.6 to match pandapower changes)
+
+    pv: ``numpy.ndarray``, dtype:np.int
+        Index of the pv buses (slack bus must NOT be on this list)
+
+    pq: ``numpy.ndarray``, dtype:np.int
+        Index of the pq buses (slack bus must NOT be on this list)
+
+    ppci: ``dict``
+        pandapower internal "ppc", ignored.
+
+    options: ``dict``
+        Dictionnary of various pandapower option. Only "max_iteration" and "tolerance_mva" are used at the moment.
+
+    Returns
+    -------
+    V: ``numpy.ndarray``, dtype:complex
+        The final complex voltage vector
+
+    converged: ``bool``
+        Whether the powerflow converged or not
+
+    iterations: ``int``
+        The number of iterations the solver performed
+
+    J: `numpy.sparmatrix``, dtype:float
+        The csc scipy matrix of the jacobian matrix of the system.
+
+    """
+    max_it = options["max_iteration"]
+    tol = options['tolerance_mva']
+    bus = ppci['bus']
+    slack_weights = bus[:, SL_FAC].astype(float)  ## contribution factors for distributed slack
+    slack_weights /= slack_weights.sum()
+
     # initialize the solver
     # TODO have that in options maybe (can use GaussSeidel, and NR with KLU -faster- or SparseLU)
     if KLU_solver_available:
@@ -80,7 +182,7 @@ def newtonpf(Ybus, Sbus, V0, pv, pq, ppci, options):
     Ybus = sparse.csc_matrix(Ybus)
 
     # do the newton raphson algorithm
-    solver.solve(Ybus, V0, Sbus, pv, pq, max_it, tol)
+    solver.solve(Ybus, V0, Sbus, ref, slack_weights, pv, pq, max_it, tol)
 
     # extract the results
     Va = solver.get_Va()