+++ src/BaseNRSolver.h
@@ -14,16 +14,21 @@
 /**
 Base class for Newton Raphson based solver
 **/
+template<class LinearSolver>
 class BaseNRSolver : public BaseSolver
 {
     public:
         BaseNRSolver():need_factorize_(true), timer_initialize_(0.), timer_dSbus_(0.), timer_fillJ_(0.) {}
 
-        ~BaseNRSolver(){}
-
         virtual
-        Eigen::SparseMatrix<real_type> get_J(){
+        Eigen::Ref<const Eigen::SparseMatrix<real_type> > get_J() const {
             return J_;
+        }
+        
+        virtual
+        Eigen::SparseMatrix<real_type> get_J_python() const {
+            Eigen::SparseMatrix<real_type> res = get_J();
+            return res;
         }
 
         virtual
@@ -47,8 +52,7 @@ class BaseNRSolver : public BaseSolver
                         real_type tol
                         ) ;
 
-        virtual
-        void reset();
+        virtual void reset();
 
     protected:
         virtual void reset_timer(){
@@ -58,10 +62,29 @@ class BaseNRSolver : public BaseSolver
             timer_initialize_ = 0.;
         }
         virtual
-        void initialize()=0;
+        void initialize(){
+            auto timer = CustTimer();
+            n_ = static_cast<int>(J_.cols()); // should be equal to J_.nrows()
+            err_ = 0; // reset error message
+            const bool init_ok = _linear_solver.initialize(J_);
+            if(!init_ok){
+                std::cout << "init_ok " << init_ok << std::endl;
+                err_ = 1;
+            }
+            need_factorize_ = false;
+            timer_initialize_ += timer.duration();
+        }
 
         virtual
-        void solve(RealVect & b, bool has_just_been_inialized)=0;
+        void solve(RealVect & b, bool has_just_been_inialized){
+            auto timer = CustTimer();
+            const int solve_status = _linear_solver.solve(J_, b, has_just_been_inialized);
+            if(solve_status != 0){
+                std::cout << "solve error: " << solve_status << std::endl;
+                err_ = solve_status;
+            }
+            timer_solve_ += timer.duration();
+        }
 
         void _dSbus_dV(const Eigen::Ref<const Eigen::SparseMatrix<cplx_type> > & Ybus,
                        const Eigen::Ref<const CplxVect > & V);
@@ -116,6 +139,8 @@ class BaseNRSolver : public BaseSolver
                             const Eigen::VectorXi & pvpq);
 
     protected:
+        // used linear solver
+        LinearSolver _linear_solver;
 
         // solution of the problem
         Eigen::SparseMatrix<real_type> J_;  // the jacobian matrix
@@ -140,7 +165,7 @@ class BaseNRSolver : public BaseSolver
                                     const Eigen::VectorXi & pq,
                                     const Eigen::VectorXi & pvpq
                                     ){
-            // DO NOT USE, FOR DEBUG ONLY!// TODO SLACK
+            // DO NOT USE, FOR DEBUG ONLY (especially for multiple slacks)
             const auto & n_pvpq = pvpq.size();
             const auto & n_pq = pvpq.size();
             std::vector<int> pvpq_inv(V.size(), -1);
@@ -195,5 +220,7 @@ class BaseNRSolver : public BaseSolver
             return res;
         }
 };
+
+#include "BaseNRSolver.tpp"
 
 #endif // BASENRSOLVER_H